ipam: Ensure the package builds on macOS
Provide a stub that fails at runtime.
